refactor: extract assistant provider adapters
