+++ docs/repository-maturity-plan.md
@@ -151,6 +151,8 @@ scripts/
 40. `assistant.js` 开始复用 assistant tool ui helper，ask_user 交互卡片与 tool progress 渲染开始从页面文件中收口。
 41. 新增 `src/lib/assistant-tool-orchestrator.js`，抽离 tool history entry 构造/收尾与等待态包装逻辑。
 42. `callAIWithTools(...)` 开始复用 assistant tool orchestrator helper，tool 调度编排继续从页面文件中收口。
+43. 新增 `src/lib/assistant-provider-adapters.js`，抽离多 provider API 调用、SSE 读取与工具定义格式转换逻辑。
+44. `assistant.js` 开始复用 assistant provider adapters helper，provider-specific 调用入口继续从页面文件中剥离。
 
 ## 风险与回滚建议
 - 风险：`chat.js` 仍然较大，后续继续拆分时容易影响事件时序。
